Splits Pavlov commands into separate files per permission level
