+++ bot/cogs/pavlov.py
@@ -11,12 +11,7 @@ import discord
 from discord.ext import commands
 
 from bot.utils import Paginator, aliases, servers
-from bot.utils.pavlov import (
-    check_perm_admin,
-    check_perm_captain,
-    check_perm_moderator,
-    exec_server_command,
-)
+from bot.utils.pavlov import exec_server_command
 from bot.utils.steamplayer import SteamPlayer
 from bot.utils.text_to_image import text_to_image
 from bs4 import BeautifulSoup
@@ -27,9 +22,6 @@ from bs4 import BeautifulSoup
 # Everyone - RefreshList, InspectPlayer, ServerInfo
 # Ban – Told to fuck off
 
-
-MATCH_DELAY_RESETSND = 10
-RCON_COMMAND_PAUSE = 100 / 1000  # milliseconds
 
 ANYONEPLAYING_ROW_FORMAT = (
     "{alias:^15} | {server_name:^36.36} | {map_name:^36.36} "
@@ -270,291 +262,6 @@ class Pavlov(commands.Cog):
         await ctx.send(embed=embed)
 
     @commands.command()
-    async def switchmap(self, ctx, map_name: str, game_mode: str, server_name: str):
-        """`{prefix}switchmap <map_name> <game_mode> <server_name>`
-
-        **Requires**: Captain permissions or higher for the server
-        **Example**: `{prefix}switchmap 89374583439127 rush`
-        """
-        if not await check_perm_captain(ctx, server_name):
-            return
-        map_label = aliases.get_map(map_name)
-        data = await exec_server_command(
-            ctx, server_name, f"SwitchMap {map_label} {game_mode}"
-        )
-        switch_map = data.get("SwitchMap")
-        if ctx.batch_exec:
-            return switch_map
-        if not switch_map:
-            embed = discord.Embed(
-                description=f"**Failed** to switch map to {map_name} with game mode {game_mode}"
-            )
-        else:
-            embed = discord.Embed(
-                description=f"Switched map to {map_name} with game mode {game_mode}"
-            )
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def resetsnd(self, ctx, server_name: str):
-        """`{prefix}resetsnd <server_name>`
-
-        **Requires**: Captain permissions or higher for the server
-        **Example**: `{prefix}resetsnd rush`
-        """
-        if not await check_perm_captain(ctx, server_name):
-            return
-        data = await exec_server_command(ctx, server_name, "ResetSND")
-        reset_snd = data.get("ResetSND")
-        if ctx.batch_exec:
-            return reset_snd
-        if not reset_snd:
-            embed = discord.Embed(description=f"**Failed** reset SND")
-        else:
-            embed = discord.Embed(description=f"SND successfully reset")
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def switchteam(self, ctx, player_arg: str, team_id: str, server_name: str):
-        """`{prefix}switchteam <player_id> <team_id> <server_name>`
-
-        **Requires**: Captain permissions or higher for the server
-        **Example**: `{prefix}resetsnd 89374583439127 0 rush`
-        """
-        if not await check_perm_captain(ctx, server_name):
-            return
-        player = SteamPlayer.convert(player_arg)
-        data = await exec_server_command(
-            ctx, server_name, f"SwitchTeam {player.unique_id} {team_id}"
-        )
-        switch_team = data.get("SwitchTeam")
-        if ctx.batch_exec:
-            return switch_team
-        if not switch_team:
-            embed = discord.Embed(
-                description=f"**Failed** to switch <{player.unique_id}> to team {team_id}"
-            )
-        else:
-            embed = discord.Embed(
-                description=f"<{player.unique_id}> switched to team {team_id}"
-            )
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def rotatemap(self, ctx, server_name: str):
-        """`{prefix}rotatemap <server_name>`
-
-        **Requires**: Captain permissions or higher for the server
-        **Example**: `{prefix}rotatemap rush`
-        """
-        if not await check_perm_captain(ctx, server_name):
-            return
-        data = await exec_server_command(ctx, server_name, f"RotateMap")
-        rotate_map = data.get("RotateMap")
-        if ctx.batch_exec:
-            return rotate_map
-        if not rotate_map:
-            embed = discord.Embed(description=f"**Failed** to rotate map")
-        else:
-            embed = discord.Embed(description=f"Rotated map successfully")
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def ban(self, ctx, player_arg: str, server_name: str):
-        """`{prefix}ban <player_id> <server_name>`
-
-        **Requires**: Moderator permissions or higher for the server
-        **Example**: `{prefix}ban 89374583439127 rush`
-        """
-        if not await check_perm_moderator(ctx, server_name):
-            return
-        player = SteamPlayer.convert(player_arg)
-        data = await exec_server_command(ctx, server_name, f"Ban {player.unique_id}")
-        ban = data.get("Ban")
-        if ctx.batch_exec:
-            return ban
-        if not ban:
-            embed = discord.Embed(description=f"**Failed** to ban <{player.unique_id}>")
-        else:
-            embed = discord.Embed(
-                description=f"<{player.unique_id}> successfully banned"
-            )
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def kill(self, ctx, player_arg: str, server_name: str):
-        """`{prefix}kill <player_id> <server_name>`
-
-        **Requires**: Moderator permissions or higher for the server
-        **Example**: `{prefix}kill 89374583439127 rush`
-        """
-        if not await check_perm_moderator(ctx, server_name):
-            return
-        player = SteamPlayer.convert(player_arg)
-        data = await exec_server_command(ctx, server_name, f"Kill {player.unique_id}")
-        kill = data.get("Kill")
-        if ctx.batch_exec:
-            return kill
-        if not kill:
-            embed = discord.Embed(
-                description=f"**Failed** to kill <{player.unique_id}>"
-            )
-        else:
-            embed = discord.Embed(
-                description=f"<{player.unique_id}> successfully killed"
-            )
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def kick(self, ctx, player_arg: str, server_name: str):
-        """`{prefix}kick <player_id> <server_name>`
-
-        **Requires**: Moderator permissions or higher for the server
-        **Example**: `{prefix}kick 89374583439127 rush`
-        """
-        if not await check_perm_moderator(ctx, server_name):
-            return
-        player = SteamPlayer.convert(player_arg)
-        data = await exec_server_command(ctx, server_name, f"Kick {player.unique_id}")
-        kick = data.get("Kick")
-        if ctx.batch_exec:
-            return kick
-        if not kick:
-            embed = discord.Embed(
-                description=f"**Failed** to kick <{player.unique_id}>"
-            )
-        else:
-            embed = discord.Embed(
-                description=f"<{player.unique_id}> successfully kicked"
-            )
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def unban(self, ctx, player_arg: str, server_name: str):
-        """`{prefix}unban <player_id> <server_name>`
-
-        **Requires**: Moderator permissions or higher for the server
-        **Example**: `{prefix}unban 89374583439127 rush`
-        """
-        if not await check_perm_moderator(ctx, server_name):
-            return
-        player = SteamPlayer.convert(player_arg)
-        data = await exec_server_command(ctx, server_name, f"Unban {player.unique_id}")
-        unban = data.get("Unban")
-        if ctx.batch_exec:
-            return unban
-        if not unban:
-            embed = discord.Embed(
-                description=f"**Failed** to unban <{player.unique_id}>"
-            )
-        else:
-            embed = discord.Embed(
-                description=f"<{player.unique_id}> successfully unbanned"
-            )
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def giveitem(self, ctx, player_arg: str, item_id: str, server_name: str):
-        """`{prefix}giveitem <player_id> <item_id> <server_name>`
-
-        **Requires**: Admin permissions for the server
-        **Example**: `{prefix}giveitem 89374583439127 tazer rush`
-        """
-        if not await check_perm_admin(ctx, server_name):
-            return
-        player = SteamPlayer.convert(player_arg)
-        data = await exec_server_command(
-            ctx, server_name, f"GiveItem {player.unique_id} {item_id}"
-        )
-        give_team = data.get("GiveItem")
-        if ctx.batch_exec:
-            return give_team
-        if not give_team:
-            embed = discord.Embed(
-                description=f"**Failed** to give {item_id} to <{player.unique_id}>"
-            )
-        else:
-            embed = discord.Embed(
-                description=f"{item_id} given to <{player.unique_id}>"
-            )
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def givecash(self, ctx, player_arg: str, cash_amount: str, server_name: str):
-        """`{prefix}givecash <player_id> <cash_amount> <server_name>`
-
-        **Requires**: Admin permissions for the server
-        **Example**: `{prefix}givecash 89374583439127 5000 rush`
-        """
-        if not await check_perm_admin(ctx, server_name):
-            return
-        player = SteamPlayer.convert(player_arg)
-        data = await exec_server_command(
-            ctx, server_name, f"GiveCash {player.unique_id} {cash_amount}"
-        )
-        give_cash = data.get("GiveCash")
-        if ctx.batch_exec:
-            return give_cash
-        if not give_cash:
-            embed = discord.Embed(
-                description=f"**Failed** to give {cash_amount} to <{player.unique_id}>"
-            )
-        else:
-            embed = discord.Embed(
-                description=f"{cash_amount} given to <{player.unique_id}>"
-            )
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def giveteamcash(self, ctx, team_id: str, cash_amount: str, server_name: str):
-        """`{prefix}giveteamcash <team_id> <cash_amount> <server_name>`
-
-        **Requires**: Admin permissions for the server
-        **Example**: `{prefix}giveteamcash 0 5000 rush`
-        """
-        if not await check_perm_admin(ctx, server_name):
-            return
-        data = await exec_server_command(
-            ctx, server_name, f"GiveTeamCash {team_id} {cash_amount}"
-        )
-        give_team_cash = data.get("GiveTeamCash")
-        if ctx.batch_exec:
-            return give_team_cash
-        if not give_team_cash:
-            embed = discord.Embed(
-                description=f"**Failed** to give {cash_amount} to <{team_id}>"
-            )
-        else:
-            embed = discord.Embed(description=f"{cash_amount} given to <{team_id}>")
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def setplayerskin(self, ctx, player_arg: str, skin_id: str, server_name: str):
-        """`{prefix}setplayerskin <player_id> <skin_id> <server_name>`
-
-        **Requires**: Admin permissions for the server
-        **Example**: `{prefix}setplayerskin 89374583439127 clown rush`
-        """
-        if not await check_perm_admin(ctx, server_name):
-            return
-        player = SteamPlayer.convert(player_arg)
-        data = await exec_server_command(
-            ctx, server_name, f"SetPlayerSkin {player.unique_id} {skin_id}"
-        )
-        set_player_skin = data.get("SetPlayerSkin")
-        if ctx.batch_exec:
-            return set_player_skin
-        if not set_player_skin:
-            embed = discord.Embed(
-                description=f"**Failed** to set <{player.unique_id}>'s skin to {skin_id}"
-            )
-        else:
-            embed = discord.Embed(
-                description=f"<{player.unique_id}>'s skin set to {skin_id}"
-            )
-        await ctx.send(embed=embed)
-
-    @commands.command()
     async def batch(self, ctx, *batch_commands):
         """`{prefix}batch "<command with arguments>" "<command with args>"`
 
@@ -590,44 +297,6 @@ class Pavlov(commands.Cog):
                     value="execution failed - command not found",
                     inline=False,
                 )
-        embed.set_footer(text=f"Execution time: {datetime.now() - before}")
-        await ctx.send(embed=embed)
-
-    @commands.command()
-    async def matchsetup(
-        self, ctx, team_a_name: str, team_b_name: str, server_name: str
-    ):
-        """`{prefix}matchsetup <CT team name> <T team name> <server name>`
-
-        **Requires**: Captain permissions or higher for the server
-        **Example**: `{prefix}matchsetup ct_team t_team rush`
-        """
-        if not await check_perm_captain(ctx, server_name):
-            return
-        before = datetime.now()
-        teams = [aliases.get_team(team_a_name), aliases.get_team(team_b_name)]
-        embed = discord.Embed()
-        for team in teams:
-            embed.add_field(
-                name=f"{team.name} members", value=team.member_repr(), inline=False
-            )
-        await ctx.send(embed=embed)
-
-        for index, team in enumerate(teams):
-            for member in team.members:
-                await exec_server_command(
-                    ctx, server_name, f"SwitchTeam {member.unique_id} {index}"
-                )
-                await asyncio.sleep(RCON_COMMAND_PAUSE)
-
-        await ctx.send(
-            embed=discord.Embed(
-                description=f"Teams set up. Resetting SND in {MATCH_DELAY_RESETSND} seconds."
-            )
-        )
-        await asyncio.sleep(MATCH_DELAY_RESETSND)
-        await exec_server_command(ctx, server_name, "ResetSND")
-        embed = discord.Embed(description="Reset SND. Good luck!")
         embed.set_footer(text=f"Execution time: {datetime.now() - before}")
         await ctx.send(embed=embed)
