+++ bot/cogs/pavlov_admin.py
@@ -1,0 +1,121 @@
+import logging
+
+import discord
+from discord.ext import commands
+
+from bot.utils import SteamPlayer
+from bot.utils.pavlov import check_perm_admin, exec_server_command
+
+
+class PavlovAdmin(commands.Cog):
+    def __init__(self, bot):
+        self.bot = bot
+
+    @commands.Cog.listener()
+    async def on_ready(self):
+        logging.info(f"{type(self).__name__} Cog ready.")
+
+    @commands.command()
+    async def giveitem(self, ctx, player_arg: str, item_id: str, server_name: str):
+        """`{prefix}giveitem <player_id> <item_id> <server_name>`
+
+        **Requires**: Admin permissions for the server
+        **Example**: `{prefix}giveitem 89374583439127 tazer rush`
+        """
+        if not await check_perm_admin(ctx, server_name):
+            return
+        player = SteamPlayer.convert(player_arg)
+        data = await exec_server_command(
+            ctx, server_name, f"GiveItem {player.unique_id} {item_id}"
+        )
+        give_team = data.get("GiveItem")
+        if ctx.batch_exec:
+            return give_team
+        if not give_team:
+            embed = discord.Embed(
+                description=f"**Failed** to give {item_id} to <{player.unique_id}>"
+            )
+        else:
+            embed = discord.Embed(
+                description=f"{item_id} given to <{player.unique_id}>"
+            )
+        await ctx.send(embed=embed)
+
+    @commands.command()
+    async def givecash(self, ctx, player_arg: str, cash_amount: str, server_name: str):
+        """`{prefix}givecash <player_id> <cash_amount> <server_name>`
+
+        **Requires**: Admin permissions for the server
+        **Example**: `{prefix}givecash 89374583439127 5000 rush`
+        """
+        if not await check_perm_admin(ctx, server_name):
+            return
+        player = SteamPlayer.convert(player_arg)
+        data = await exec_server_command(
+            ctx, server_name, f"GiveCash {player.unique_id} {cash_amount}"
+        )
+        give_cash = data.get("GiveCash")
+        if ctx.batch_exec:
+            return give_cash
+        if not give_cash:
+            embed = discord.Embed(
+                description=f"**Failed** to give {cash_amount} to <{player.unique_id}>"
+            )
+        else:
+            embed = discord.Embed(
+                description=f"{cash_amount} given to <{player.unique_id}>"
+            )
+        await ctx.send(embed=embed)
+
+    @commands.command()
+    async def giveteamcash(self, ctx, team_id: str, cash_amount: str, server_name: str):
+        """`{prefix}giveteamcash <team_id> <cash_amount> <server_name>`
+
+        **Requires**: Admin permissions for the server
+        **Example**: `{prefix}giveteamcash 0 5000 rush`
+        """
+        if not await check_perm_admin(ctx, server_name):
+            return
+        data = await exec_server_command(
+            ctx, server_name, f"GiveTeamCash {team_id} {cash_amount}"
+        )
+        give_team_cash = data.get("GiveTeamCash")
+        if ctx.batch_exec:
+            return give_team_cash
+        if not give_team_cash:
+            embed = discord.Embed(
+                description=f"**Failed** to give {cash_amount} to <{team_id}>"
+            )
+        else:
+            embed = discord.Embed(description=f"{cash_amount} given to <{team_id}>")
+        await ctx.send(embed=embed)
+
+    @commands.command()
+    async def setplayerskin(self, ctx, player_arg: str, skin_id: str, server_name: str):
+        """`{prefix}setplayerskin <player_id> <skin_id> <server_name>`
+
+        **Requires**: Admin permissions for the server
+        **Example**: `{prefix}setplayerskin 89374583439127 clown rush`
+        """
+        if not await check_perm_admin(ctx, server_name):
+            return
+        player = SteamPlayer.convert(player_arg)
+        data = await exec_server_command(
+            ctx, server_name, f"SetPlayerSkin {player.unique_id} {skin_id}"
+        )
+        set_player_skin = data.get("SetPlayerSkin")
+        if ctx.batch_exec:
+            return set_player_skin
+        if not set_player_skin:
+            embed = discord.Embed(
+                description=f"**Failed** to set <{player.unique_id}>'s skin to {skin_id}"
+            )
+        else:
+            embed = discord.Embed(
+                description=f"<{player.unique_id}>'s skin set to {skin_id}"
+            )
+        await ctx.send(embed=embed)
+
+
+def setup(bot):
+    bot.add_cog(PavlovAdmin(bot))
